requirements.txt
aiofiles==23.1.0
aioredis==2.0.1
anyio==3.6.2
astroid==2.14.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
dill==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
orjson==3.8.7
passlib==1.7.4
platformdirs==3.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pylint==2.16.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.15.0
